Visualization aids to utils.py, plt.rcParams.update to init.py
